pyproject.toml
[project]
name = "tagihan-wifi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.0.0",
    "duckdb>=1.4.1",
    "email-validator>=2.1.0",
    "fastapi>=0.121.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "sqids>=0.5.2",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "autoimport>=1.6.1",
    "icecream>=2.1.8",
    "isort>=7.0.0",
]
